fixed link error, added sub class, and more
Fixed a link error that existed in the Animals page, added a sub class
using the Page class as a super(this made a great template) we can now
pass things to the sub class, and even override the super class.
Altered Main.py so it now uses the sub class rather than the super.